# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for the spv-wallet-web-frontend.",
        "title": "SPV Wallet WEB Backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/transaction": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create transaction.",
                "parameters": [
                    {
                        "description": "Create transaction data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transports_http_endpoints_api_transactions.CreateTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transports_http_endpoints_api_transactions.FullTransaction"
                        }
                    }
                }
            }
        },
...
        "/api/v1/transaction/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get transaction by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transports_http_endpoints_api_transactions.FullTransaction"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "transports_http_endpoints_api_transactions.CreateTransaction": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "satoshis": {
                    "type": "integer"
                }
            }
        },
        "transports_http_endpoints_api_transactions.FullTransaction": {
            "type": "object",
            "properties": {
                "blockHash": {
                    "type": "string"
                },
                "blockHeight": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "fee": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "numberOfInputs": {
                    "type": "integer"
                },
                "numberOfOutputs": {
                    "type": "integer"
                },
                "receiver": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "totalValue": {
                    "type": "integer"
                }
            }
        },
...
    }
}